Fixed a small bug.
Added logic in lines 41 - 47 to make sure that the dividend that is shown in the spreadsheet is the dividend that will be next paid.  For example, some companies may declare their Q4 dividend before the Q3 dividend is paid.  Without this fix, the pay date and ex dividend date (and dividend amount) would show as the Q4 dividend.